Rename file and make cache read safer
